pyproject.toml
[tool.poetry]
name = "rec"
version = "0.1.0"
description = ""
authors = ["Alex <sashoism@pm.me>"]

[tool.poetry.dependencies]
python = "^3.10"
pydub = "^0.25.1"
pulsectl-asyncio = "^0.2.3"
python-socketio = {extras = ["asyncio_client"], version = "^5.7.1"}
toml = "^0.10.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
